dashboard_control: carve api_media.rs (pure move)
